Separate the connected components into a different folder
